Added Hint for "A" Cell Recomposition Device
A good friend of mine always complains about this card not asking you in the End Phase, so I guess you can say he's the reason I'm adding the hint. :D

Hints for these cards are a little more needed, due to it also having, "...except the turn this card was sent to the Graveyard...",
so if you forgot to hold the 'A' button, you now have a hint to remind you so you don't miss out on its search effect.

I added quite a few hints recently, where they're not really needed, so if you disagree with adding them, kindly notify me and I'll close this request to allow other people's requests to advance. Also, if someone can teach me how to "squash" (combine) all these pull requests into one, that would be helpful.