fix EFFECT_DISABLE_FIELD
EFFECT_DISABLE_FIELD has 3 types:
1. The player chooses the location
The script has an operation function but no value function.
The temp value will be set in operation, and it will be reset in card::reset().

2. Certain location related to the card
The script has a value function.
It will not be reset in card::reset().

3. Fixed location
The script sets the value as a constant.
The temp value will be reset in card::reset().