update continuous cards, part 2
The continuous spell/trap cards has this property:
"The effect resolution will do nothing if the card is no longer on the field."
It is checked by the flag CHAIN_CONTINUOUS_CARD in field::solve_chain().

It is a property of continuous spell/trap cards, not the effect itself.
When a monster copies that effect, the monster effect will not have this property.
Therefore, it should NOT be written in the script.