Fix attempt for Sphere Mode
Currently Sphere Mode does not return to the owner if it's facedown during the End Phase of the next turn after being summoned. The effect should still reset if Sphere Mode were to leave the field/temp banished, but not if facedown. I dont know if theres a better way to fix this, but this at least seems to work.

https://db.ygoresources.com/card#11927

"Even if this card that was Summoned to your opponent's field is in face-down Defense Position, the process of "control of this card returns to its original owner at the End Phase of the next turn" will proceed as usual."