Fix Magnet Warrior Σ+
Fix an issue where the controller of Magnet Warrior Σ+ was allowed to choose the opponents attack targets while controlling 2 or more EARTH monsters, while the opponent controlled none.

It appears to be an issue with (EFFECT_MUST_ATTACK_MONSTER) which is granting the controller the ability to choose the attack targets for the opponent, so long as theyre still attacking EARTH monsters. Ive used Nightmare Pain as a reference for this fix, which appears to work fine after testing.